Fixed ticket names

+++ src/main/java/AfkoAPI/DAO/AbbreviationDao.java
@@ -86,9 +86,9 @@ public class AbbreviationDao {
     public HTTPResponse getAbbreviationByNameOrOrgId(String name, String orgId, String amount) {
         List<Abbreviation> data = null;
         if (orgId.equals(""))
-            data = abbrRep.findByNameStartsWith(name);
+            data = abbrRep.findByNameStartsWithIgnoreCase(name);
         else
-             data = abbrRep.findByNameStartsWithAndOrganisations_id(name, orgId);
+             data = abbrRep.findByNameStartsWithIgnoreCaseAndOrganisations_id(name, orgId);
 
         TrimListService<Abbreviation> service = new TrimListService<>();
         data = service.trimListByAmount(data, amount);
@@ -158,11 +158,13 @@ public class AbbreviationDao {
      * @param abbrs the abbreviations to remove
      * @return an HTTPResponse containing the given abbreviations
      */
-    public HTTPResponse deleteAbbreviations(Abbreviation[] abbrs){
+    public HTTPResponse<Abbreviation[]> deleteAbbreviations(Abbreviation[] abbrs){
         for (Abbreviation abbr: abbrs){
+            Optional<Abbreviation> a = abbrRep.findById(abbr.getId());
+            if (a.isEmpty()) return HTTPResponse.<Abbreviation[]>returnFailure("could not find abbreviation with id: " + abbr.getId());
             abbrRep.deleteById(abbr.getId());
         }
-        return HTTPResponse.returnSuccess(abbrs);
+        return HTTPResponse.<Abbreviation[]>returnSuccess(abbrs);
     }
 }
